flask_server/requirements.txt
Flask==2.0.1
Pillow==8.3.2
Werkzeug==2.0.1 